Fix WSConnection leaks
Fixes issue #14.

There were leaks at two levels: the R6 object and the underlying
C++ WSConnection object. Both problems originated in the bindings
to the handlers (the arguments to set_XXX_handler). The first was
that the Rcpp::Functions being passed to bind() were holding refs
to the R6 object by default. Moving them onto the WSConnection and
clearing them during close() fixes this, and the R6 is successfully
collected.

For the WSConnection object, the passing of the shared_ptr to bind
created a refcount cycle. Using weak_ptr instead breaks the cycle.